Test0, Stats 96-99
Healthbar ui, daño elemental, hay que hacer el 100

+++ Assets/2-Scripts/Stat.cs
@@ -19,7 +19,7 @@ public class Stat
         return finalValue;
     }
 
-    public void SetDefaultValue(int value)
+    public void SetDefaultValue(float value)
     {
         baseValue = value;
     }
